feat: generate command rework
The subcommand `cloud-account iac-generate` is now top level `generate` command
