+++ integration/test_resources/help/generate
@@ -1,19 +1,16 @@
-Manage cloud account integrations with Lacework
+Create IaC content for various different cloud environments and configurations
 
 Usage:
-  lacework cloud-account [command]
+  lacework generate [command]
 
 Aliases:
-  cloud-account, cloud-accounts, cloud, ca
+  generate, gen
 
 Available Commands:
-  create      Create a new cloud account integration
-  delete      Delete a cloud account integration
-  list        List all available cloud account integrations
-  show        Show a single cloud account integration
+  cloud-account Create IaC code
 
 Flags:
-  -h, --help   help for cloud-account
+  -h, --help   help for generate
 
 Global Flags:
   -a, --account string      account subdomain of URL (i.e. <ACCOUNT>.lacework.net)
@@ -29,4 +26,4 @@ Global Flags:
   -p, --profile string      switch between profiles configured at ~/.lacework.toml
       --subaccount string   sub-account name inside your organization (org admins only)
 
-Use "lacework cloud-account [command] --help" for more information about a command.
+Use "lacework generate [command] --help" for more information about a command.
