+++ integration/test_resources/help/generate_cloud-account
@@ -1,10 +1,10 @@
-Create IaC content for various different cloud environments and configurations
+Create IaC content for various different cloud environments
 
 Usage:
-  lacework cloud-account iac-generate [command]
+  lacework generate cloud-account [command]
 
 Aliases:
-  iac-generate, iac
+  cloud-account, cloud, ca
 
 Available Commands:
   aws         Generate and/or execute Terraform code for AWS integration
@@ -12,7 +12,7 @@ Available Commands:
   gcp         Generate and/or execute Terraform code for GCP integration
 
 Flags:
-  -h, --help   help for iac-generate
+  -h, --help   help for cloud-account
 
 Global Flags:
   -a, --account string      account subdomain of URL (i.e. <ACCOUNT>.lacework.net)
@@ -28,4 +28,4 @@ Global Flags:
   -p, --profile string      switch between profiles configured at ~/.lacework.toml
       --subaccount string   sub-account name inside your organization (org admins only)
 
-Use "lacework cloud-account iac-generate [command] --help" for more information about a command.
+Use "lacework generate cloud-account [command] --help" for more information about a command.
