+++ integration/test_resources/help/generate_cloud-account_gcp
@@ -16,7 +16,7 @@ This command can also be run in noninteractive mode.
 See help output for more details on the parameter value(s) required for Terraform code generation.
 
 Usage:
-  lacework cloud-account iac-generate gcp [flags]
+  lacework generate cloud-account gcp [flags]
 
 Flags:
       --apply                                         run terraform apply without executing plan or prompting
